// apps/jobs/web/src/lib/api/generated/model/jobInputApplyMode.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire Jobs API
 * Jobs mini-app API
 * OpenAPI spec version: 0.1.0
 */

export type JobInputApplyMode = typeof JobInputApplyMode[keyof typeof JobInputApplyMode];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const JobInputApplyMode = {
  internal: 'internal',
  external: 'external',
} as const;