// apps/jobs/web/src/lib/api/generated/model/jobInputSalaryPeriod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire Jobs API
 * Jobs mini-app API
 * OpenAPI spec version: 0.1.0
 */

export type JobInputSalaryPeriod = typeof JobInputSalaryPeriod[keyof typeof JobInputSalaryPeriod];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const JobInputSalaryPeriod = {
  year: 'year',
  month: 'month',
  hour: 'hour',
  project: 'project',
} as const;