// apps/jobs/web/src/lib/api/generated/model/jobInput.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire Jobs API
 * Jobs mini-app API
 * OpenAPI spec version: 0.1.0
 */
import type { JobInputFunction } from './jobInputFunction';
import type { JobInputSeniority } from './jobInputSeniority';
import type { JobInputEmploymentType } from './jobInputEmploymentType';
import type { JobInputWorkplace } from './jobInputWorkplace';
import type { JobInputLocation } from './jobInputLocation';
import type { JobInputSalary } from './jobInputSalary';
import type { JobInputApplyMode } from './jobInputApplyMode';
import type { JobInputScreeningQuestionsItem } from './jobInputScreeningQuestionsItem';

export interface JobInput {
  /**
   * @minLength 3
   * @maxLength 120
   */
  title: string;
  /**
   * @minLength 20
   * @maxLength 12000
   */
  description: string;
  function: JobInputFunction;
  seniority: JobInputSeniority;
  employmentType: JobInputEmploymentType;
  workplace: JobInputWorkplace;
  location?: JobInputLocation;
  salary?: JobInputSalary;
  /** @maxItems 15 */
  skills?: string[];
  applyMode?: JobInputApplyMode;
  /** @maxLength 500 */
  externalApplyUrl?: string;
  /** @maxItems 5 */
  screeningQuestions?: JobInputScreeningQuestionsItem[];
}